move editable files to imports bye bye /client/widgets
