+++ client/main.scss
@@ -29,7 +29,7 @@
 @import '{}/imports/ui/stylesheets/rtl';
 
 // Template styles
-@import 'widgets/editable/editable';
+@import '{}/imports/ui/components/editable/editable';
 @import '{}/imports/ui/components/map/map';
 @import '{}/imports/ui/components/messages/messages';
 @import '{}/imports/ui/components/region-tag/region-tag';
