+++ server/extracts.msgfmt~
@@ -1,1 +1,1 @@
-[{"map.addMarker":{"key":"map.addMarker","text":"set marker","file":"imports/ui/components/map/map.js","line":123,"mtime":1509277291230,"template":"FaIcon","ctime":1509277291230},"map.fullscreen":{"key":"map.fullscreen","text":"big map","file":"imports/ui/components/map/map.js","line":113,"mtime":1509277291230,"template":"FaIcon","ctime":1509277291230},"map.fullscreenClose":{"key":"map.fullscreenClose","text":"close","file":"imports/ui/components/map/map.js","line":118,"mtime":1509277291230,"template":"FaIcon","ctime":1509277291230},"map.removeMarker":{"key":"map.removeMarker","text":"remove the marker","file":"imports/ui/components/map/map.js","line":128,"mtime":1509277291230,"template":"FaIcon","ctime":1509277291230},"map.zoomInTitle":{"key":"map.zoomInTitle","text":"zoom in","file":"imports/ui/components/map/map.js","line":103,"mtime":1509277291230,"template":"FaIcon","ctime":1509277291230},"map.zoomOutTitle":{"key":"map.zoomOutTitle","text":"zoom out","file":"imports/ui/components/map/map.js","line":104,"mtime":1509277291230,"template":"FaIcon","ctime":1509277291230}},{"extractedAt":1509277303176,"updatedAt":1509277291230}]
+[{"course.detail.remove":{"_id":"P7zfHMT3NuxrgFTCo","key":"course.detail.remove","lang":"en","text":"Remove course and all its events?","ctime":null,"mtime":1498474350522,"revisionId":"hCuLTLsux5tLpiDBi","file":"client/views/courses/details/course.details.js","line":136,"template":"unknown","removed":true},"editable.add_text":{"key":"editable.add_text","text":"Add text here","file":"imports/ui/lib/editable.js","line":81,"mtime":1509294693053,"template":"unknown","ctime":1509294693053}},{"extractedAt":1509294723372,"updatedAt":1509294693053}]
